src: Move trivia generation code to separate class from visitor.
