+++ src/PortToTripleSlash/src/libraries/Docs/DocsCommentsContainer.cs
@@ -1,4 +1,4 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
@@ -13,7 +13,7 @@ namespace ApiDocsSync.PortToTripleSlash.Docs
 {
     internal class DocsCommentsContainer
     {
-        private Configuration Config { get; set; }
+        internal Configuration Config { get; }
 
         public readonly Dictionary<string, DocsType> Types = new();
         public readonly Dictionary<string, DocsMember> Members = new();
